app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dotenv==0.9.9
fastapi==0.128.5
feedparser==6.0.12
greenlet==3.3.1
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sgmllib3k==1.0.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
